tests/test_robots_handler.py
import unittest
from unittest.mock import patch, MagicMock
from robots_handler import fetch_and_store_robots, check_robots_rules
import mysql.connector


class DummyRobotFileParser:
    def __init__(self):
        self.disallow_all = ["/private"]
        self.allow_all = ["/public"]
        self.delay = 5

    def set_url(self, url):
        self.url = url

    def read(self):
        pass

    def crawl_delay(self, user_agent):
        return self.delay


class DummyCursor:
    def __init__(self):
        self.executed = []

    def execute(self, sql, params):
        self.executed.append((sql, params))

    def close(self):
        pass


class DummyConnection:
    def __init__(self):
        self.cursor_obj = DummyCursor()

    def cursor(self):
        return self.cursor_obj

    def commit(self):
        pass

    def close(self):
        pass


class TestRobotsHandler(unittest.TestCase):
    @patch("robots_handler.get_connection")
    @patch("robots_handler.RobotFileParser")
    def test_fetch_and_store_robots(self, mock_rfp_class, mock_connect):
        # 正常系: robots.txt取得・保存
        mock_rfp = MagicMock()
        mock_rfp.read.return_value = None
        mock_rfp.crawl_delay.return_value = 5
        mock_rfp.disallow_all = ["/private"]
        mock_rfp.allow_all = ["/public"]
        mock_rfp_class.return_value = mock_rfp

        mock_conn = MagicMock()
        mock_cursor = MagicMock()
        mock_connect.return_value = mock_conn
        mock_conn.cursor.return_value = mock_cursor

        fetch_and_store_robots("example.com", "MyScraperBot")

        mock_rfp.read.assert_called()
        mock_cursor.execute.assert_called()
        mock_conn.commit.assert_called()

    @patch("robots_handler.get_connection")
    @patch("robots_handler.fetch_and_store_robots")
    def test_check_robots_rules_allow(self, mock_fetch, mock_connect):
        mock_conn = MagicMock()
        mock_cursor = MagicMock()
        mock_connect.return_value = mock_conn
        mock_conn.cursor.return_value = mock_cursor

        # Allow case
        mock_cursor.fetchone.side_effect = [
            {"disallow": "/private", "allow": "/public", "crawl_delay": 3}
        ]
        result, delay = check_robots_rules(
            "https://example.com/public/page", "MyScraperBot"
        )
        self.assertTrue(result)
        self.assertEqual(delay, 3)

        # Disallow case
        mock_cursor.fetchone.side_effect = [
            {"disallow": "/private", "allow": "/public", "crawl_delay": 3}
        ]
        result, delay = check_robots_rules(
            "https://example.com/private/data", "MyScraperBot"
        )
        self.assertFalse(result)
        self.assertEqual(delay, 0)

        # crawl_delay unset case
        mock_cursor.fetchone.side_effect = [
            {"disallow": "/private", "allow": "/public", "crawl_delay": None}
        ]
        result, delay = check_robots_rules(
            "https://example.com/public/page", "MyScraperBot"
        )
        self.assertTrue(result)
        self.assertEqual(delay, 0)
        mock_conn = MagicMock()
        mock_cursor = MagicMock()
        mock_connect.return_value = mock_conn
        mock_conn.cursor.return_value = mock_cursor

        mock_cursor.fetchone.side_effect = [
            {"disallow": "/private", "allow": "/public", "crawl_delay": 3}
        ]

        result, delay = check_robots_rules(
            "https://example.com/public/page", "MyScraperBot"
        )
        self.assertTrue(result)
        self.assertEqual(delay, 3)

    def test_fetch_and_store_robots_monkeypatch(self):
        # Monkey patch RobotFileParser using unittest.mock.patch
        with patch(
            "robots_handler.RobotFileParser", new=lambda: DummyRobotFileParser()
        ):
            # Monkey patch mysql.connector.connect using unittest.mock.patch
            with patch(
                "mysql.connector.connect", new=lambda **kwargs: DummyConnection()
            ):
                # Execute function
                domain = "example.com"
                fetch_and_store_robots(domain)

                # Instantiate new connection to verify dummy behavior
                conn = mysql.connector.connect()
                cursor = conn.cursor()
                # Use the dummy RobotFileParser to check crawl_delay functionality
                rp = DummyRobotFileParser()
                rp.read()  # No-op for dummy
                delay = rp.crawl_delay("MyScraperBot")
                self.assertEqual(delay, 5)
        # Monkey patch RobotFileParser and mysql.connector.connect
        #  using unittest.mock.patch
        with patch(
            "robots_handler.RobotFileParser", new=lambda: DummyRobotFileParser()
        ):
            with patch(
                "mysql.connector.connect", new=lambda **kwargs: DummyConnection()
            ):
                # Execute function
                domain = "example.com"
                fetch_and_store_robots(domain)

                # Retrieve the executed SQL from the dummy connection
                conn = mysql.connector.connect()
                cursor = conn.cursor()
                # Dummy execute to populate executed list
                cursor.execute("SELECT ...", ())
                # Since the function uses its own connection instance,
                #  we cannot retrieve that here directly,
                # so instead, we test that our dummy RobotFileParser works as expected
                rp = DummyRobotFileParser()
                rp.read()
                delay = rp.crawl_delay("MyScraperBot")
                self.assertEqual(delay, 5)

    def dummy_connect_fail(*args, **kwargs):
        raise Exception("Database connection failed")

    def test_fetch_and_store_robots_db_error(self):
        from unittest.mock import patch

        # Use dummy RobotFileParser for consistency
        with patch(
            "robots_handler.RobotFileParser", new=lambda: DummyRobotFileParser()
        ), patch("mysql.connector.connect", side_effect=self.dummy_connect_fail):
            with self.assertRaises(Exception) as context:
                fetch_and_store_robots("example.com")
        self.assertIn("Database connection failed", str(context.exception))
